Exceptions Added

+++ OnlineFoodDeliverySystem/Serivces/CustomerService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using OnlineFoodDeliverySystem.Data;
+using OnlineFoodDeliverySystem.Exceptions;
 using OnlineFoodDeliverySystem.Repository;
 
 namespace OnlineFoodDeliverySystem.Services
@@ -16,26 +17,48 @@ namespace OnlineFoodDeliverySystem.Services
         }
         public async Task<IEnumerable<Customer>> GetAllCustomersAsync()
         {
-            return await _customerRepository.GetAllCustomersAsync();
+            var customers = await _customerRepository.GetAllCustomersAsync();
+            return customers.ToList();
         }
 
         public async Task<Customer> GetCustomerByIdAsync(int customerId)
         {
-            return await _customerRepository.GetCustomerByIdAsync(customerId);
+            Customer customer = await _customerRepository.GetCustomerByIdAsync(customerId);
+            if (customer == null)
+            {
+                throw new NotFoundException($"Customer with id {customerId} does not exists");
+            }
+            return customer;
         }
 
         public async Task AddCustomerAsync(Customer customer)
         {
+            var Addcustomer = await _customerRepository.GetCustomerByIdAsync(customer.CustomerID);
+            if(Addcustomer != null)
+            {
+                throw new AlreadyExistsException($"Customer with {customer.CustomerID} already exists");
+            }
             await _customerRepository.AddCustomerAsync(customer);
+
         }
 
         public async Task UpdateCustomerAsync(Customer customer)
         {
+            var Updatecustomer = await _customerRepository.GetCustomerByIdAsync(customer.CustomerID);
+            if(Updatecustomer == null)
+            {
+                throw new NotFoundException($"Customer with {customer.CustomerID} does not exists");
+            }
             await _customerRepository.UpdateCustomerAsync(customer);
         }
 
         public async Task DeleteCustomerAsync(int customerId)
         {
+            Customer Deletecustomer = await _customerRepository.GetCustomerByIdAsync(customerId);
+            if (Deletecustomer == null)
+            {
+                throw new NotFoundException($"Customer with id {customerId} does not exists");
+            }
             await _customerRepository.DeleteCustomerAsync(customerId);
         }
     }
